Upgrade MySql.Data
There is a bug in version 6 of the MySql.Data client library that throws an exception when the server has multiple IP addresses. This bug has been fixed in version 8. As this is only used in the tests and we do not take a direct dependency on it, it is safe to upgrade it.